deleted AI comments and capitalized LED
