+++ src/main/java/frc/robot/constants/LEDConstants.java
@@ -1,0 +1,14 @@
+package frc.robot.constants;
+
+public class LEDConstants {
+    public static final int LED_PWM_PORT = 9;
+    public static final int LED_STRIP_LENGTH = 60;
+
+    public static final int ANIMATION_SPEED = 3;
+    public static final int PATTERN_LENGTH = 20;
+
+    public static final int[] COLOR_BLUE = {0, 0, 255};
+    public static final int[] COLOR_YELLOW = {255, 255, 0};
+    public static final int[] COLOR_GREEN = {0, 255, 0};
+    public static final int[] COLOR_RED = {255, 0, 0};
+}
